requirements/dev.txt
aiosqlite==0.21.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
base58==2.1.1
bcrypt==4.3.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coincurve==21.0.0
colorama==0.4.6
cytoolz==1.0.1
ecdsa==0.19.1
eth-hash==0.7.1
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.12
flake8==7.1.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
parsimonious==0.10.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pyasn1==0.4.8
pycodestyle==2.12.1
pycryptodome==3.22.0
pydantic==2.11.0
pydantic-settings==2.8.1
pydantic_core==2.33.0
pyflakes==3.2.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
rsa==4.9
ruff==0.11.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
toolz==1.0.0
tronpy==0.5.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
